Migrate two controls into one
